+++ CHANGES.md
@@ -1,4 +1,12 @@
-# SMART_MRS (2024-01-21)
+# SMART_MRS (2025-04-14)
+**Release Highlights**
+SMART_MRS (PyPi version 2.1) has the following updates:
+- Minor improvements to artifact functions
+    - Spurious echo updated to T2 vs. T_all
+    - Larmor frequency now in Hz for ALL relevant equations
+    - Corrected inconsistencies in phase component implementation in relevant equations
+
+# SMART_MRS (2025-01-21)
 **Release Highlights**
 SMART_MRS (PyPi version 2) has the following updates:
 - Improvements to artifact functions